requirements.txt
aiogram~=3.13.1
APScheduler~=3.10.4
python-decouple~=3.8
aiohttp~=3.10.8
redis~=5.1.0
asyncpg~=0.29.0